Add Rijksmuseum to multiplatform-samples
Hallo, Rijksmuseum is Kotlin and Compose Multiplatform app that offers an immersive way to explore the art collection of the renowned Rijksmuseum in Amsterdam. The project is multi-modular and uses compose for all platforms (Android, iOS, desktop and web). it also showcases topics like DI, Networking, Navigation and Caching...

If this is a good sample for people, please include it in the doc, or else feel free to close the PR.

Here is the repo: https://github.com/fethij/rijksmuseum