update: multiplatform resources page restructured
We separated overview of the library, the setup page, and the actual use cases page.

The section on `composeResources {}` settings block was divided, to only mention relevant settings where it is appropriate right now. If there will be a lot of settings in the future, we can compile a reference on a separate page.

Also added are some EAP-related changes, mostly taken from the already published What's new in 1.7.0 page.